more cleanups and cast removal

+++ include/msl/msl_types.h
@@ -5,7 +5,7 @@
  * Public ILP32 MSL facade types used by game code. Sound playback returns a
  * list-pool ID, not an object pointer; zero is the invalid handle.
  */
-typedef unsigned long MslSoundHandle;
+typedef unsigned int MslSoundHandle;
 
 typedef struct _mslSystem _mslSystem;
 typedef struct mslLoadedBank mslLoadedBank;
